Moving refactored mini functions from utils to controllers folder
